Split the management tests by version instead of by machinery
test_versioned_management.py grew around the cross-version bridge, so it
mixed three unrelated things: version-neutral helper units, v1 entity
behavior, and structural checks on the split itself. With the bridge gone
the file has no subject. Replace it with four files, each targeting exactly
one version, and no test branching on version:

  test_management.py           v1 -- workspaces and workspace groups
  test_management_v2.py        v2 -- the flat Cluster resource
  test_management_utils.py     version-neutral helper units
  test_management_versioning.py structural invariants of the split

The version-neutral units now import the shared modules they actually
exercise (management.files, management.stage, management.organization)
rather than reaching through v1/ for the same objects.

TestV1IsDeletable becomes TestVersionPackagesAreIndependent and enforces
rule 1 in both directions: the AST walk and the sys.meta_path import
blocker each now run v2->v1 and v1->v2. It also asserts the inheritance
direction (shared base <- version subclass, never v1 <- v2).

Adds coverage for the Part 4 deprecation: manage_workspaces() warns and
_manage_workspaces_v1() stays silent under -W error::DeprecationWarning.

test_management_v2.py has NOT been run against a live v2 organization; it
was written by translating the v1 suites resource by resource. Every
assertion resting on a v2 request or response shape rather than on
SDK-internal behavior carries an UNVERIFIED comment -- create_cluster's
POST body most of all, where the nested size object and the
provider/region pair replacing v1's regionID are both unconfirmed.
