+++ singlestoredb/tests/test_management_utils.py
@@ -1,0 +1,467 @@
+#!/usr/bin/env python
+# type: ignore
+"""
+Version-neutral unit tests for the management API helpers.
+
+Nothing here touches a version-specific module or needs a management token or
+a container. These were originally written alongside the versioned wrappers
+only because that is where the bugs were found.
+"""
+import datetime
+import os
+import pathlib
+import unittest
+from unittest.mock import MagicMock
+
+from singlestoredb.exceptions import ManagementError
+from singlestoredb.management.utils import normalize_remote_path
+
+
+TEST_DIR = pathlib.Path(os.path.dirname(__file__))
+
+
+class TestFolderTransferPaths(unittest.TestCase):
+    """Folder helpers must address remote objects with the full remote path
+    and resolve ``ignore`` globs relative to the local folder."""
+
+    def _make_stage(self):
+        from singlestoredb.management.stage import Stage
+        stage = Stage.__new__(Stage)
+        stage._manager = MagicMock()
+        return stage
+
+    def _make_file_space(self):
+        from singlestoredb.management.files import FileSpace
+        space = FileSpace.__new__(FileSpace)
+        space._manager = MagicMock()
+        return space
+
+    def _make_files_object(self, path, type_='file'):
+        from singlestoredb.management.files import FilesObject
+        return FilesObject(
+            name=path.rsplit('/', 1)[-1],
+            path=path,
+            size=0,
+            type=type_,
+            format='',
+            mimetype='',
+            created=None,
+            last_modified=None,
+            writable=True,
+        )
+
+    def _make_local_tree(self, tmp):
+        """Create ``<tmp>/src/keep.py`` and ``<tmp>/src/sub/skip.pyc``."""
+        import os
+        root = os.path.join(tmp, 'src')
+        os.makedirs(os.path.join(root, 'sub'))
+        keep = os.path.join(root, 'keep.py')
+        skip = os.path.join(root, 'sub', 'skip.pyc')
+        for path in (keep, skip):
+            with open(path, 'w') as f:
+                f.write('x')
+        return root, keep, skip
+
+    def test_stage_download_folder_prefixes_remote_paths(self):
+        import tempfile
+        stage = self._make_stage()
+        # listdir strips the stage_path prefix from its results
+        stage.listdir = MagicMock(
+            return_value=[
+                self._make_files_object('a.txt'),
+                self._make_files_object('sub/b.txt'),
+            ],
+        )
+        stage.is_dir = MagicMock(side_effect=lambda p: p == 'remote/folder')
+        stage._download_file = MagicMock()
+        with tempfile.TemporaryDirectory() as tmp:
+            stage.download_folder('remote/folder', tmp, overwrite=True)
+        requested = [call.args[0] for call in stage._download_file.call_args_list]
+        self.assertEqual(
+            requested, ['remote/folder/a.txt', 'remote/folder/sub/b.txt'],
+        )
+
+    def test_stage_download_folder_normalizes_prefix(self):
+        import tempfile
+        stage = self._make_stage()
+        stage.listdir = MagicMock(
+            return_value=[self._make_files_object('a.txt')],
+        )
+        # download_folder normalizes './remote/folder/' before probing.
+        stage.is_dir = MagicMock(side_effect=lambda p: p == 'remote/folder')
+        stage._download_file = MagicMock()
+        with tempfile.TemporaryDirectory() as tmp:
+            stage.download_folder('./remote/folder/', tmp, overwrite=True)
+        self.assertEqual(
+            stage._download_file.call_args_list[0].args[0],
+            'remote/folder/a.txt',
+        )
+
+    def test_stage_download_folder_uses_listing_type_not_is_dir(self):
+        """The entry type comes from the listing, so no per-entry is_dir
+        call is made, and empty remote folders are still created locally."""
+        import os
+        import tempfile
+        stage = self._make_stage()
+        stage.listdir = MagicMock(
+            return_value=[
+                self._make_files_object('empty', type_='directory'),
+                self._make_files_object('a.txt'),
+            ],
+        )
+        is_dir_calls = []
+
+        def is_dir(p):
+            is_dir_calls.append(p)
+            return p == 'remote'
+
+        stage.is_dir = is_dir
+        stage._download_file = MagicMock()
+        with tempfile.TemporaryDirectory() as tmp:
+            dest = os.path.join(tmp, 'dest')
+            stage.download_folder('remote', dest, overwrite=True)
+            # Only the top-level folder check, nothing per entry
+            self.assertEqual(is_dir_calls, ['remote'])
+            self.assertTrue(os.path.isdir(os.path.join(dest, 'empty')))
+        requested = [call.args[0] for call in stage._download_file.call_args_list]
+        self.assertEqual(requested, ['remote/a.txt'])
+
+    def test_stage_upload_folder_ignores_folder_patterns(self):
+        import os
+        import tempfile
+        stage = self._make_stage()
+        stage.exists = MagicMock(return_value=False)
+        stage.upload_file = MagicMock()
+        stage.info = MagicMock()
+        with tempfile.TemporaryDirectory() as tmp:
+            root = os.path.join(tmp, 'src')
+            os.makedirs(os.path.join(root, '__pycache__'))
+            keep = os.path.join(root, 'keep.py')
+            for path in (keep, os.path.join(root, '__pycache__', 'a.pyc')):
+                with open(path, 'w') as f:
+                    f.write('x')
+            stage.upload_folder(root, 'dest', ignore='**/__pycache__')
+            uploaded = [
+                call.args[0] for call in stage.upload_file.call_args_list
+            ]
+            self.assertEqual(uploaded, [keep])
+
+    def test_file_space_upload_folder_ignores_folder_patterns(self):
+        import os
+        import tempfile
+        space = self._make_file_space()
+        space.upload_file = MagicMock()
+        space.info = MagicMock()
+        with tempfile.TemporaryDirectory() as tmp:
+            root = os.path.join(tmp, 'src')
+            os.makedirs(os.path.join(root, '__pycache__'))
+            keep = os.path.join(root, 'keep.py')
+            for path in (keep, os.path.join(root, '__pycache__', 'a.pyc')):
+                with open(path, 'w') as f:
+                    f.write('x')
+            space.upload_folder(root, 'dest', ignore='**/__pycache__')
+            uploaded = [
+                call.kwargs['local_path']
+                for call in space.upload_file.call_args_list
+            ]
+            self.assertEqual(uploaded, [keep])
+
+    def test_download_folder_defaults_to_remote_folder_name(self):
+        """With no local_path, the destination is the remote folder's name
+        in the current directory."""
+        import os
+        import tempfile
+        cwd = os.getcwd()
+        for name, obj, attr in (
+            ('Stage', self._make_stage(), '_download_file'),
+            ('FileSpace', self._make_file_space(), '_download_file'),
+        ):
+            obj.listdir = MagicMock(
+                return_value=[self._make_files_object('a.txt')],
+            )
+            obj.is_dir = MagicMock(return_value=True)
+            setattr(obj, attr, MagicMock())
+            with tempfile.TemporaryDirectory() as tmp:
+                try:
+                    os.chdir(tmp)
+                    obj.download_folder('remote/folder')
+                finally:
+                    os.chdir(cwd)
+                target = getattr(obj, attr).call_args_list[0].args[1]
+                self.assertEqual(
+                    os.path.normpath(target),
+                    os.path.join('folder', 'a.txt'),
+                    f'{name} wrote to {target}',
+                )
+
+    def test_download_folder_root_without_local_path_raises(self):
+        for obj in (self._make_stage(), self._make_file_space()):
+            obj.listdir = MagicMock(return_value=[])
+            obj.is_dir = MagicMock(return_value=True)
+            with self.assertRaises(ValueError) as ctx:
+                obj.download_folder('/')
+            self.assertIn('local_path must be specified', str(ctx.exception))
+
+    def test_download_folder_explicit_local_path_unchanged(self):
+        """Explicit local_path keeps writing directly into that directory."""
+        import os
+        import tempfile
+        for obj in (self._make_stage(), self._make_file_space()):
+            obj.listdir = MagicMock(
+                return_value=[self._make_files_object('a.txt')],
+            )
+            obj.is_dir = MagicMock(return_value=True)
+            obj._download_file = MagicMock()
+            with tempfile.TemporaryDirectory() as tmp:
+                dest = os.path.join(tmp, 'dest')
+                obj.download_folder('remote/folder', dest, overwrite=True)
+                self.assertEqual(
+                    obj._download_file.call_args_list[0].args[1],
+                    os.path.join(dest, 'a.txt'),
+                )
+
+    def test_upload_folder_builds_slash_separated_remote_paths(self):
+        """Remote paths must use '/' even when the local platform uses '\\'."""
+        import tempfile
+        stage = self._make_stage()
+        stage.exists = MagicMock(return_value=False)
+        stage.upload_file = MagicMock()
+        stage.info = MagicMock()
+        with tempfile.TemporaryDirectory() as tmp:
+            root, _, _ = self._make_local_tree(tmp)
+            stage.upload_folder(root, 'dest/')
+            targets = sorted(
+                call.args[1] for call in stage.upload_file.call_args_list
+            )
+            self.assertEqual(targets, ['dest/keep.py', 'dest/sub/skip.pyc'])
+            for target in targets:
+                self.assertNotIn('\\', target)
+
+    def test_stage_upload_folder_applies_ignore_globs(self):
+        import tempfile
+        stage = self._make_stage()
+        stage.exists = MagicMock(return_value=False)
+        stage.upload_file = MagicMock()
+        stage.info = MagicMock()
+        with tempfile.TemporaryDirectory() as tmp:
+            root, keep, _ = self._make_local_tree(tmp)
+            stage.upload_folder(root, 'dest', ignore='**/*.pyc')
+            uploaded = [
+                call.args[0] for call in stage.upload_file.call_args_list
+            ]
+            self.assertEqual(uploaded, [keep])
+
+    def test_stage_upload_folder_applies_ignore_globs_to_cwd(self):
+        import os
+        import tempfile
+        stage = self._make_stage()
+        stage.exists = MagicMock(return_value=False)
+        stage.upload_file = MagicMock()
+        stage.info = MagicMock()
+        cwd = os.getcwd()
+        with tempfile.TemporaryDirectory() as tmp:
+            root, _, _ = self._make_local_tree(tmp)
+            try:
+                os.chdir(root)
+                stage.upload_folder('.', 'dest', ignore='**/*.pyc')
+            finally:
+                os.chdir(cwd)
+            uploaded = [
+                call.args[0] for call in stage.upload_file.call_args_list
+            ]
+            self.assertEqual(uploaded, ['keep.py'])
+
+    def test_file_space_upload_folder_applies_ignore_globs(self):
+        import tempfile
+        space = self._make_file_space()
+        space.upload_file = MagicMock()
+        space.info = MagicMock()
+        with tempfile.TemporaryDirectory() as tmp:
+            root, keep, _ = self._make_local_tree(tmp)
+            space.upload_folder(root, 'dest', ignore='**/*.pyc')
+            uploaded = [
+                call.kwargs['local_path']
+                for call in space.upload_file.call_args_list
+            ]
+            self.assertEqual(uploaded, [keep])
+
+    def test_file_space_upload_folder_applies_ignore_globs_to_cwd(self):
+        import os
+        import tempfile
+        space = self._make_file_space()
+        space.upload_file = MagicMock()
+        space.info = MagicMock()
+        cwd = os.getcwd()
+        with tempfile.TemporaryDirectory() as tmp:
+            root, _, _ = self._make_local_tree(tmp)
+            try:
+                os.chdir(root)
+                space.upload_folder('.', 'dest', ignore='**/*.pyc')
+            finally:
+                os.chdir(cwd)
+            uploaded = [
+                call.kwargs['local_path']
+                for call in space.upload_file.call_args_list
+            ]
+            self.assertEqual(uploaded, ['keep.py'])
+
+
+class TestRecursiveDownloadPathTraversal(unittest.TestCase):
+    """Recursive download helpers must refuse to write outside ``local_path``
+    when the remote listing contains traversal segments (``..``)."""
+
+    def _make_file_location(self):
+        # FileSpace is a concrete FileLocation subclass; instantiate via
+        # __new__ to skip its constructor (which expects a real FilesManager).
+        from singlestoredb.management.files import FileSpace
+        loc = FileSpace.__new__(FileSpace)
+        loc._manager = MagicMock()
+        return loc
+
+    def _make_files_object(self, path, type_='file'):
+        from singlestoredb.management.files import FilesObject
+        return FilesObject(
+            name=path.rsplit('/', 1)[-1],
+            path=path,
+            size=0,
+            type=type_,
+            format='',
+            mimetype='',
+            created=None,
+            last_modified=None,
+            writable=True,
+        )
+
+    def test_files_download_folder_rejects_traversal(self):
+        import tempfile
+        loc = self._make_file_location()
+        # Listing returns an entry whose path escapes via '..'
+        loc.listdir = MagicMock(
+            return_value=[self._make_files_object('../escape.txt')],
+        )
+        loc._download_file = MagicMock()
+        with tempfile.TemporaryDirectory() as tmp:
+            target = f'{tmp}/dest'
+            import os
+            os.makedirs(target)
+            with self.assertRaises(ManagementError) as ctx:
+                loc.download_folder('remote', target, overwrite=True)
+            self.assertIn('outside destination', str(ctx.exception))
+            loc._download_file.assert_not_called()
+
+    def test_files_download_folder_rejects_traversal_directory(self):
+        import tempfile
+        loc = self._make_file_location()
+        # Directory entry that escapes
+        loc.listdir = MagicMock(
+            return_value=[self._make_files_object('../evil', type_='directory')],
+        )
+        with tempfile.TemporaryDirectory() as tmp:
+            target = f'{tmp}/dest'
+            import os
+            os.makedirs(target)
+            with self.assertRaises(ManagementError) as ctx:
+                loc.download_folder('remote', target, overwrite=True)
+            self.assertIn('outside destination', str(ctx.exception))
+
+    def test_stage_download_folder_rejects_traversal(self):
+        import tempfile
+        from singlestoredb.management.stage import Stage
+        stage = Stage.__new__(Stage)
+        stage.listdir = MagicMock(
+            return_value=[self._make_files_object('../escape.txt')],
+        )
+        # is_dir(stage_path) must return True (it's a directory); the entry
+        # type in the listing marks each entry as a file.
+        stage.is_dir = MagicMock(side_effect=lambda p: p == 'remote')
+        stage._download_file = MagicMock()
+        with tempfile.TemporaryDirectory() as tmp:
+            target = f'{tmp}/dest'
+            import os
+            os.makedirs(target)
+            with self.assertRaises(ManagementError) as ctx:
+                stage.download_folder('remote', target, overwrite=True)
+            self.assertIn('outside destination', str(ctx.exception))
+            stage._download_file.assert_not_called()
+
+
+class TestRemotePathUtils(unittest.TestCase):
+    """Test cases for remote path normalization (no server required)."""
+
+    def test_local_separators_converted(self):
+        # A prefix built with os.path.join on Windows keeps a trailing '\'
+        assert normalize_remote_path('llama3\\') == 'llama3'
+        assert normalize_remote_path('a\\b\\c.txt') == 'a/b/c.txt'
+        assert normalize_remote_path(pathlib.PurePosixPath('a/b')) == 'a/b'
+
+    def test_duplicate_and_trailing_separators_collapsed(self):
+        assert normalize_remote_path('a//b/') == 'a/b'
+        assert normalize_remote_path('a/b///') == 'a/b'
+        assert normalize_remote_path('a\\\\b\\') == 'a/b'
+
+    def test_strip_leading(self):
+        assert normalize_remote_path('./a/b', strip_leading=True) == 'a/b'
+        assert normalize_remote_path('/a/b', strip_leading=True) == 'a/b'
+        assert normalize_remote_path('.\\a\\b', strip_leading=True) == 'a/b'
+        assert normalize_remote_path('/', strip_leading=True) == ''
+        assert normalize_remote_path('', strip_leading=True) == ''
+
+    def test_strip_leading_off_by_default(self):
+        assert normalize_remote_path('/a/b') == '/a/b'
+
+    def test_joining_produces_valid_remote_path(self):
+        # Regression: 'llama3\/file' was produced before normalization
+        prefix = normalize_remote_path('llama3\\')
+        assert f'{prefix}/file' == 'llama3/file'
+
+    def test_listdir_style_suffix(self):
+        # The listdir call sites append '/' after normalizing
+        assert normalize_remote_path('llama3\\', strip_leading=True) + '/' \
+            == 'llama3/'
+        assert normalize_remote_path('/', strip_leading=True) + '/' == '/'
+
+
+class TestSecretFromDictTimestamps(unittest.TestCase):
+    """
+    Coverage for ``Secret.from_dict`` running its timestamp fields
+    through ``to_datetime``.
+    """
+
+    def test_timestamps_parsed_to_datetime(self):
+        from singlestoredb.management.organization import Secret
+
+        obj = {
+            'secretID': 'sec-1',
+            'name': 'my-secret',
+            'createdBy': 'user-a',
+            'createdAt': '2024-01-01T00:00:00Z',
+            'lastUpdatedBy': 'user-b',
+            'lastUpdatedAt': '2024-02-15T12:34:56Z',
+            'value': 'shh',
+            'deletedBy': None,
+            'deletedAt': None,
+        }
+        sec = Secret.from_dict(obj)
+        self.assertIsInstance(sec.created_at, datetime.datetime)
+        self.assertEqual(sec.created_at.year, 2024)
+        self.assertIsInstance(sec.last_updated_at, datetime.datetime)
+        self.assertEqual(sec.last_updated_at.minute, 34)
+        self.assertIsNone(sec.deleted_at)
+
+    def test_missing_timestamps_become_none(self):
+        from singlestoredb.management.organization import Secret
+
+        obj = {
+            'secretID': 'sec-1',
+            'name': 'my-secret',
+            'createdBy': 'user-a',
+            'lastUpdatedBy': 'user-b',
+        }
+        sec = Secret.from_dict(obj)
+        self.assertIsNone(sec.created_at)
+        self.assertIsNone(sec.last_updated_at)
+        self.assertIsNone(sec.deleted_at)
+
+
+if __name__ == '__main__':
+    unittest.main()
